Pytorch/极客时间学习/实战-语义分割-UNet/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
Pillow>=8.0.0
tqdm>=4.60.0

